requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
glicko2==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
livereload==2.7.1
MarkupSafe==3.0.3
six==1.17.0
SQLAlchemy==2.0.48
tornado==6.5.5
trueskill==0.4.5
typing_extensions==4.15.0
Werkzeug==3.1.7
